Restructure feature flags
