+++ netwatch/Cargo.toml
@@ -41,8 +41,8 @@ noq-udp = "0.9"
 libc = "0.2.139"
 socket2 = { version = "0.6", features = ["all"] }
 
-# netdev is only available on desktop/mobile platforms
-[target.'cfg(any(target_os = "linux", target_os = "android", target_os = "macos", target_os = "ios", target_os = "windows", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))'.dependencies]
+# netdev is available on all non-embedded, non-wasm platforms
+[target.'cfg(not(any(target_os = "espidf", all(target_family = "wasm", target_os = "unknown"))))'.dependencies]
 netdev = "0.40.1"
 tokio = { version = "1", features = [
     "fs",
