+++ netwatch/src/lib.rs
@@ -1,9 +1,9 @@
 //! Networking related utilities
 
 #[cfg_attr(wasm_browser, path = "interfaces/wasm_browser.rs")]
-#[cfg_attr(not(any(has_netdev, wasm_browser)), path = "interfaces/fallback.rs")]
+#[cfg_attr(posix_minimal, path = "interfaces/posix_minimal.rs")]
 pub mod interfaces;
-#[cfg_attr(not(any(has_netdev, wasm_browser)), path = "ip_fallback.rs")]
+#[cfg_attr(posix_minimal, path = "ip_posix_minimal.rs")]
 pub mod ip;
 mod ip_family;
 pub mod netmon;
