+++ netwatch/src/netmon.rs
@@ -18,12 +18,12 @@ mod android;
 mod bsd;
 #[cfg(target_os = "linux")]
 mod linux;
+#[cfg(posix_minimal)]
+mod posix_minimal;
 #[cfg(wasm_browser)]
 mod wasm_browser;
 #[cfg(target_os = "windows")]
 mod windows;
-#[cfg(not(any(has_netdev, wasm_browser)))]
-mod fallback;
 
 #[cfg(not(wasm_browser))]
 pub(crate) use self::actor::is_interesting_interface;
